+++ src/lte/test/test-lte-x2-handover.cc
@@ -238,7 +238,7 @@ LteX2HandoverTestCase::DoRun ()
   Ptr<ListPositionAllocator> positionAlloc = CreateObject<ListPositionAllocator> ();
   positionAlloc->Add (Vector (-3000, 0, 0)); // enb0
   positionAlloc->Add (Vector ( 3000, 0, 0)); // enb1
-  for (uint16_t i = 0; i < m_nUes; i++)
+  for (uint32_t i = 0; i < m_nUes; i++)
     {
       positionAlloc->Add (Vector (-3000, 100, 0));
     }
